Refactor the Kaleidoscope.ino file to prepare for more program logic to handle the different 'modes' that we will need to be in.
Create a new Kaleidoscope class to encapsulate the display of the Kaleidoscope.
This required me to rename the Kaleidoscope.ino file as behind the scenes,
they must generate a Kaleidoscope.h header file which conflicted with
my new class header file.

Create a new LEDStrips class to encapsulate the LED strips. Note, this
would have been better as a singleton class but the Arduino tools apparently
don't support that.