Update KaleidoscopeLookupTable as it had a few bugs.
Switch to using CRGB instead of uint32_t anywhere possible so we can
leverage all the HSV and color helper functions in FastLED.

Rename NUM_STRIPS and NUM_LEDS_PER_STRIP to match better with FastLED
'standards' to make porting effects over easier.

Update some of the demos to make them 'kaleidoscopic'

Remove RTC clock as it is buggy (hangs after a few hours) and will be
replaces with the teensy library.

Remove seven-segment.h until it is updated