Big changes this commit. The code has been updated to run on the Teensy 4.0.
I'm using VisualTeensy to generate a project for VSCode and tycommander
to do the flash update.

Not everything is ported over yet. Still need to do the photoresister,
clock, and rotary encoders.

Don't make changes to files in the 'lib' directory - they are copied
over by VisualTeensy from the original source so any changes will be
lost the next time the libraries are updated.