Add REST API to enable controlling the Kaleidoscope from external applications. The API is as follows:
Get the current kaleidoscope settings:
------------------------------------------------
"/api/settings", HTTP_GET

"brightness" = The manual offset for the LED brightness.
"speed"  = The number of milliseconds to delay between frames.
"mode" = The name of the current Kaleidoscope mode.
"clockFace" = The name of the current clock face.

An example is:
{"brightness":3995,"speed":550,"mode":"xy_fire","clockFace":"Digital"}

Set the kaleidoscope settings:
------------------------------------------------
"/api/settings", HTTP_PUT

"brightness" = The manual offset for the LED brightness. This is constrained between -4095 and +4095
"speed"  = The number of milliseconds to delay between frames. This is constrained between 0 and 1000ms
"mode" = the name of the Kaleidoscope mode. This is constrained to those listed in a call to /api/modes, HTTP_GET
"clockFace" = the name of the clock face. This is constrained to those listed in a call to /api/faces, HTTP_GET

Return an array of available kaleidoscope modes:
------------------------------------------------
"/api/modes", HTTP_GET

An example is:
["kaleidoscope","off","kaleidoscope_select_speed_brightness","kaleidoscope_select_reflection_style","select_clock_face","off","kaleidoscope_beatWave","kaleidoscope_blendWave","kaleidoscope_plasma","kaleidoscope_rainbowMarch","kaleidoscope_ripples","kaleidoscope_twinkle_fox","xy_distortion_waves","xy_matrix","xy_pacifica","xy_rainbow","xy_fire","off"]

Return an array of available clock faces:
------------------------------------------------
"/api/faces", HTTP_GET

An example is:
["None","Digital","Analog"]