Fix the crashing bug that sent us into the config portal on every boot. AsyncDNSServer scope needs to be globa as it is apparently used long after ESPAsync_WiFiManager is done.
Add debug settings to enable JTAG debugging via the ESP-Prog. While it works, mostly,
it is slow and tends to crash a lot. :(

Clean up DEBUG code in several palces so that we can build and run with DEBUG turned on.
Unfortunately, we can't do this globally via platform.ini as the OS/WiFi/??? stack is unstable when DEBUG is on.