When WiFi is actively transmitting/receiving data, the FastLED.show() command tends to crash or hang the main 'loop()' thread. This is because the FastLed library disables interrupts during calls to show() so that it can ensure the timing is correct while updating the LED strips.
To fix this correctly requires us to switch to a different type of LED
that doesn't require such tight timing loops to update or to switch to a
different processor that has better hardware than the ESP32. Neither of
these options are great as they both require a redesigned PCB so I'll
try to minimze the conflict using a semaphore.

While this will solve the conflicts for the WiFi calls that originate
from the Kaleidoscope, it doesn't help those coming from external
clients as we don't have any control over when those arrive.