Update to current wspressif32 (3.5.0 --> 6.8.1) Update to latest versions of dependencies Enable separate release, debug, and JTAG builds
Pull user settings (mode, speed, drawStyle, brightness, clockface, clockColor)
into a settings structure so that they can be saved/restore to/from persistent
storage. The user feature this provides is that when the power is lost/restored,
the kaleidoscope will go back to the settings before power loss.

Change how the web UI is provided. Instead of encoding it in a header file,
use the SPIFFS file system and have the webserver pull the files directly
from there. This is a much simpler way to make all required files available
to the webserver.

Remove the WiFiSemaphore as it wasn't fully implemented and didn't prevent
the WiFi crashes as hoped. Turn on FASTLED_ESP32_FLASH_LOCK to force
flash operations to wait until the call to FastLED.show() is done so that
loading the WebUI page or reading/writing settings doesn't crash the system.

NOTE: While we provide async REST APIs that can change the settings on core 0 as well
as the physical knob controls that can change them manually on core 1, this shouldn't
happen often so there is currently no semaphore protection to make all settings access
thread safe.

The hostname (SSID) and timezone are read from settings but there is currently
no way to change these from the defaults.

Replace DynamicJsonDocument with JsonDocument as it has been depreciated.